pyproject.toml
[project]
name = "tobias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.12.0",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.1.3",
    "openai>=2.30.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.0",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
    "streamlit>=1.32.0",
]
